requirements_updated.txt
streamlit
pandas
numpy
tensorflow-cpu==2.13.0
scikit-learn
